backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
rapidfuzz
python-dotenv
email-validator
httpx
